+++ intelmq/bots/parsers/shadowserver/parser.py
@@ -62,6 +62,11 @@ class ShadowserverParserBot(ParserBot):
 
         # Set config to parse report
         self.report_name = report.get('extra.file_name')
+        if not self.report_name:
+            raise ValueError("No feedname given as parameter and the "
+                             "processed report has no 'extra.file_name'. "
+                             "Ensure that at least one is given. "
+                             "Also have a look at the documentation of the bot.")
         filename_search = self.__is_filename_regex.search(self.report_name)
 
         if not filename_search: